--- packages/e2e-tests/utils/noteContent.ts
@@ -1,10 +1,0 @@
-export const noteContent = `Mr. Smith is a 53-year-old male with a past medical history significant for hypertension and type 2 diabetes mellitus. He reports that approximately 3 weeks ago, he began experiencing mild shortness of breath while climbing stairs. Over the last 10 days, the symptoms have progressed, and he now becomes winded after walking less than one block. He describes occasional chest tightness, lasting about 2–3 minutes, which is not consistently related to activity. He denies frank chest pain but does endorse mild dizziness on exertion. No syncope, palpitations, or orthopnea. He has noted mild ankle swelling in the evenings.
-
-Patient denies fever, cough, recent upper respiratory infections, or hemoptysis. Appetite is unchanged, but he reports some fatigue and reduced energy. He has not been adherent with his antihypertensive medications for the past several weeks due to running out of prescriptions.
-
-Assessment.
-Mr. Smith is a 53-year-old male with a past medical history significant for hypertension and type 2 diabetes mellitus. He reports that approximately 3 weeks ago, he began experiencing mild shortness of breath while climbing stairs. Over the last 10 days, the symptoms have progressed, and he now becomes winded after walking less than one block. He describes occasional chest tightness, lasting about 2–3 minutes, which is not consistently related to activity. He denies frank chest pain but does endorse mild dizziness on exertion. No syncope, palpitations, or orthopnea. He has noted mild ankle swelling in the evenings.
-
-Patient denies fever, cough, recent upper respiratory infections, or hemoptysis. Appetite is unchanged, but he reports some fatigue and reduced energy. He has not been adherent with his antihypertensive medications for the past several weeks due to running out of prescriptions.
-
-Assessment`;
